+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/renderer/ProjectRendererI.java
@@ -1,0 +1,4 @@
+package net.splitcells.website.server.renderer;
+
+public class ProjectRendererI {
+}